requirements.txt
streamlit
google-cloud-vertexai
google-auth
google-auth-oauthlib
google-auth-httplib2